Plotly lazy loading and code splitting
